pyproject.toml
[tool.poetry]
name = "genstudio"
version = "0.1.0"
description = ""
authors = ["Matthew Huebert <me@matt.is>"]
license = "MIT"
readme = "README.md"
packages = [{include = "genstudio", from = "src"}]
include = ["**/*_build.js"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
anywidget = "^0.9.10"
traitlets = "^5.14.3"
numpy = "^1.26.4"
jax = "^0.4.24"
penzai = "^0.1.1"
html2image = "^2.0.4.3"
pillow = "^10.4.0"
orjson = "^3.10.6"

[tool.poetry.group.dev.dependencies]
jaxlib = "^0.4.24"
ipykernel = "^6.29.4"
watchfiles = "^0.21.0"
genjax = {version = "0.3.1.post258.dev0+b103dee7", source = "gcp"}
jupyter-client = "^8.6.2"
nbformat = "^5.10.4"
mkdocs = "^1.6.0"
mkdocs-jupyter = "^0.24.8"
mkdocs-terminal = "^4.4.0"
mkdocs-material = "^9.5.31"

[[tool.poetry.source]]
name = "gcp"
url = "https://us-west1-python.pkg.dev/probcomp-caliban/probcomp/simple/"
priority = "explicit"

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src", "tests"]
exclude = ["notebooks"]
typeCheckingMode = "standard"
reportUnusedExpression = false
reportFunctionMemberAccess = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
